[registry] renamed type and methods
